vote/requirements.txt
Flask
gunicorn
kafka-python
prometheus-client
jaeger-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp-proto-http
